feat(frontend): integrate admin panel into dashboard
