+++ frontend/src/lib/index.ts
@@ -1,3 +1,4 @@
 // place files you want to import through the `$lib` alias in this folder.
 export { default as MarkdownEditor } from './MarkdownEditor.svelte';
 export { default as FileTree } from './FileTree.svelte';
+export { default as AdminPanel } from './AdminPanel.svelte';
